wip separate old-bash hack and shellcheck disable
Add clarifying hyphens -- I'd thought for a while this was an
old hack for Bash empty arrays (and was puzzling over just what
it was hacking around), until I found 81194eef4 which explains
it's a hack for *old Bash* for empty arrays.

This also avoids the comment looking like it was about the
shellcheck disable, which I thought it was until I looked up the
shellcheck warning in question.  (In fact the shellcheck warning
is about how we're splitting `$configureFlags`.)