gbsplay: fix configure flags
This caused none of these flags to have any effect.  That's because
the configure command looked like this:

    ./configure --prefix=/nix/store/svhl0fjdj1jl2sqcppy5vnzpfi4gj3d3-gbsplay-2016-12-17 \
        --without-test\ --without-contrib\ --disable-devdsp\ --enable-pulse\ --disable-alsa\ --disable-midi\ --disable-nas\ --disable-dsound\ --disable-i18n

with one giant flag '--without-test --without-contrib...', containing
spaces.

This can be seen in `nix log nixpkgs.gbsplay`, in this line:

    configure flags: --prefix=/nix/store/svhl0fjdj1jl2sqcppy5vnzpfi4gj3d3-gbsplay-2016-12-17 --without-test\ --without-contrib\ --disable-devdsp\ --enable-pulse\ --disable-alsa\ --disable-midi\ --disable-nas\ --disable-dsound\ --disable-i18n

and then in the fact that features like "devdsp" and "midi" are listed
as enabled in later output, and source files like plugout_midi.c are
included in the build.

I don't have a real opinion on whether it's better to have these flags
or not, but it's clear the author's intention was to pass them.  So,
fix the attr name so they get passed.